requirements.txt
asgiref==3.8.1
Django==5.0.3
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pytesseract==0.3.10
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
